Move TeamSyncService to a separate package
